+++ go.mod
@@ -3,11 +3,18 @@ module github.com/nats-io/nats.go
 go 1.19
 
 require (
+	github.com/golang/protobuf v1.4.2
+	github.com/nats-io/nats-server/v2 v2.9.16
 	github.com/nats-io/nkeys v0.4.4
 	github.com/nats-io/nuid v1.0.1
+	google.golang.org/protobuf v1.23.0
 )
 
 require (
-	golang.org/x/crypto v0.6.0 // indirect
-	golang.org/x/sys v0.5.0 // indirect
+	github.com/klauspost/compress v1.16.4 // indirect
+	github.com/minio/highwayhash v1.0.2 // indirect
+	github.com/nats-io/jwt/v2 v2.4.1 // indirect
+	golang.org/x/crypto v0.8.0 // indirect
+	golang.org/x/sys v0.7.0 // indirect
+	golang.org/x/time v0.3.0 // indirect
 )
